+++ test/Makefile.am
@@ -27,9 +27,9 @@ all_la_unit_tests = \
 	unit-wopi-async-slow.la \
 	unit-wopi-stuck-save.la \
 	unit-insert-delete.la \
+	unit-wopi-fail-upload.la \
 	unit-httpws.la \
 	unit-password-protected.la \
-	unit-wopi-fail-upload.la \
 	unit-wopi-ownertermination.la \
 	unit-crash.la \
 	unit-wopi-save-on-exit.la \
@@ -43,7 +43,6 @@ all_la_unit_tests = \
 	unit-copy-paste.la \
 	unit-http.la \
 	unit-wopi-temp.la \
-	unit-wopi-async-upload-close.la \
 	unit-wopi.la \
 	unit-storage.la \
 	unit-wopi-saveas-with-encoded-file-name.la \
@@ -202,8 +201,6 @@ unit_oauth_la_SOURCES = UnitOAuth.cpp
 unit_oauth_la_LIBADD = $(CPPUNIT_LIBS)
 unit_wopi_la_SOURCES = UnitWOPI.cpp
 unit_wopi_la_LIBADD = $(CPPUNIT_LIBS)
-unit_wopi_async_upload_close_la_SOURCES = UnitWOPIAsyncUpload_Close.cpp
-unit_wopi_async_upload_close_la_LIBADD = $(CPPUNIT_LIBS)
 unit_wopi_async_upload_modifyclose_la_SOURCES = UnitWOPIAsyncUpload_ModifyClose.cpp
 unit_wopi_async_upload_modifyclose_la_LIBADD = $(CPPUNIT_LIBS)
 unit_wopi_async_slow_la_SOURCES = UnitWOPISlow.cpp
